requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.3
Flask-API==1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pycurl==7.43.0
pygobject==3.20.0
PyMySQL==0.9.3
python-apt==1.1.0b1+ubuntu0.16.4.4
requests==2.22.0
SQLAlchemy==1.3.4
unattended-upgrades==0.1
urllib3==1.25.3
Werkzeug==0.15.4